requirements.txt
asgiref==3.7.2
attrs==23.2.0
Automat==22.10.0
soupsieve==2.5
beautifulsoup4==4.12.3
certifi==2023.11.17
cffi==1.16.0
charset-normalizer==3.3.2
constantly==23.10.4
cryptography==42.0.2
cssselect==1.2.0
Django==4.2.9
django-lifecycle==1.1.2
django-taggit==5.0.1
filelock==3.13.1
hyperlink==21.0.0
idna==3.6
incremental==22.10.0
itemadapter==0.8.0
itemloaders==1.1.0
jmespath==1.0.1
lxml==5.1.0
packaging==23.2
parsel==1.8.1
pillow==10.2.0
Protego==0.3.0
pyasn1==0.5.1
pyasn1-modules==0.3.0
pycparser==2.21
PyDispatcher==2.0.7
pyOpenSSL==24.0.0
python-dotenv==1.0.1
queuelib==1.6.2
requests==2.31.0
requests-file==2.0.0
Scrapy==2.11.0
selenium==4.18.0
service-identity==24.1.0
six==1.16.0
sqlparse==0.4.4
tldextract==5.1.1
Twisted==22.10.0
typing_extensions==4.9.0
urllib3==2.2.0
webdriver-manager==4.0.1
w3lib==2.1.2
zope.interface==6.1
dj-database-url==2.1.0
djangorestframework==3.15.2
psycopg2-binary==2.9.9
algoliasearch-django==3.0.0
whitenoise==6.7.0
gunicorn==22.0.0
django-lifecycle==1.2.4
pydantic==2.9.2
pycountry==24.6.1
huey
redis

